added the right favicon
