+++ agenta-web/src/pages/_document.tsx
@@ -3,7 +3,10 @@ import {Html, Head, Main, NextScript} from "next/document"
 export default function Document() {
     return (
         <Html lang="en">
-            <Head />
+            <Head>
+                <title>Agenta: The LLMOps platform.</title>
+                <link rel="shortcut icon" href="/assets/favicon.ico" />
+            </Head>
             <body>
                 <Main />
                 <NextScript />
